Fix race condition, cleanup appearance settings, misc
