+++ src/features/settings/appearance/comments/Comments.tsx
@@ -1,0 +1,15 @@
+import { IonLabel, IonList } from "@ionic/react";
+import CollapsedByDefault, { ListHeader } from "./CollapsedByDefault";
+
+export default function Comments() {
+  return (
+    <>
+      <ListHeader>
+        <IonLabel>Comments</IonLabel>
+      </ListHeader>
+      <IonList inset>
+        <CollapsedByDefault />
+      </IonList>
+    </>
+  );
+}
