+++ src/features/settings/appearance/posts/Posts.tsx
@@ -1,0 +1,18 @@
+import { IonLabel, IonList } from "@ionic/react";
+import { ListHeader } from "../TextSize";
+import BlurNsfw from "./BlurNsfw";
+import PostSize from "./PostSize";
+
+export default function Posts() {
+  return (
+    <>
+      <ListHeader>
+        <IonLabel>Posts</IonLabel>
+      </ListHeader>
+      <IonList inset>
+        <PostSize />
+        <BlurNsfw />
+      </IonList>
+    </>
+  );
+}
