+++ src/features/settings/appearance/system/DarkMode.tsx
@@ -1,16 +1,8 @@
-import styled from "@emotion/styled";
 import { IonLabel, IonList, IonToggle } from "@ionic/react";
-import { InsetIonItem } from "../../../pages/profile/ProfileFeedItemsPage";
-import { useAppDispatch, useAppSelector } from "../../../store";
-import { setUseSystemDarkMode } from "./appearanceSlice";
+import { InsetIonItem } from "../../../../pages/profile/ProfileFeedItemsPage";
+import { useAppDispatch, useAppSelector } from "../../../../store";
+import { setUseSystemDarkMode } from "../appearanceSlice";
 import UserDarkMode from "./UserDarkMode";
-
-const ListHeader = styled.div`
-  font-size: 0.8em;
-  margin: 32px 0 -8px 32px;
-  text-transform: uppercase;
-  color: var(--ion-color-medium);
-`;
 
 export default function DarkMode() {
   const dispatch = useAppDispatch();
@@ -20,9 +12,6 @@ export default function DarkMode() {
 
   return (
     <>
-      <ListHeader>
-        <IonLabel>System</IonLabel>
-      </ListHeader>
       <IonList inset>
         <InsetIonItem>
           <IonLabel>Use System Light/Dark Mode</IonLabel>
