+++ src/features/settings/appearance/system/System.tsx
@@ -1,0 +1,14 @@
+import { IonLabel } from "@ionic/react";
+import DarkMode from "./DarkMode";
+import { ListHeader } from "../TextSize";
+
+export default function System() {
+  return (
+    <>
+      <ListHeader>
+        <IonLabel>System</IonLabel>
+      </ListHeader>
+      <DarkMode />
+    </>
+  );
+}
