+++ src/pages/settings/AppearancePage.tsx
@@ -7,10 +7,7 @@ import {
   IonToolbar,
 } from "@ionic/react";
 import AppContent from "../../features/shared/AppContent";
-import TextSize from "../../features/settings/appearance/TextSize";
-import PostView from "../../features/settings/appearance/PostView";
-import DarkMode from "../../features/settings/appearance/DarkMode";
-import CollapsedByDefault from "../../features/settings/appearance/CollapsedByDefault";
+import AppearanceSettings from "../../features/settings/appearance/AppearanceSettings";
 
 export default function AppearancePage() {
   return (
@@ -25,10 +22,7 @@ export default function AppearancePage() {
         </IonToolbar>
       </IonHeader>
       <AppContent scrollY>
-        <TextSize />
-        <CollapsedByDefault />
-        <PostView />
-        <DarkMode />
+        <AppearanceSettings />
       </AppContent>
     </IonPage>
   );
